requirements.txt
openai==0.27.7
langchain==0.0.174
pypdf==3.8.1
chromadb==0.3.23
tiktoken==0.4.0
python-dotenv==1.0.0
urllib3==1.26.6